+++ Sources/XcodeGenCLI/GenerationError.swift
@@ -7,6 +7,7 @@ import Rainbow
 enum GenerationError: Error, CustomStringConvertible, ProcessError {
     case missingProjectSpec(Path)
     case projectSpecParsingError(Error)
+    case cacheGenerationError(Error)
     case validationError(SpecValidationError)
     case generationError(Error)
     case writingError(Error)
@@ -17,6 +18,8 @@ enum GenerationError: Error, CustomStringConvertible, ProcessError {
             return "No project spec found at \(path.absolute())"
         case .projectSpecParsingError(let error):
             return "Parsing project spec failed: \(error)"
+        case .cacheGenerationError(let error):
+            return "Couldn't generate cache file: \(error)"
         case .validationError(let error):
             return error.description
         case .generationError(let error):