+++ Sources/XcodeGenKit/PathExtensions.swift
@@ -3,7 +3,7 @@ import PathKit
 
 extension Path {
 
-    func byRemovingBase(path: Path) -> Path {
+    public func byRemovingBase(path: Path) -> Path {
         return Path(normalize().string.replacingOccurrences(of: "\(path.normalize().string)/", with: ""))
     }
 }